refactor(text-input-react): refactor for readability
affects: @fremtind/jkl-text-input-react
