+++ packages/text-input-react/src/TextArea.tsx
@@ -1,9 +1,10 @@
-import React, { forwardRef, FocusEvent, useEffect, useRef, useState, RefObject } from "react";
+import React, { forwardRef, FocusEvent, useRef, useState, RefObject } from "react";
+import classNames from "classnames";
+import nanoid from "nanoid";
 import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
 import { LabelVariant } from "@fremtind/jkl-core";
 import { BaseProps } from "./BaseInputField";
-import classNames from "classnames";
-import nanoid from "nanoid";
+import { useExpandingTextArea } from "./useExpandingTextArea";
 
 interface Props extends BaseProps {
     label: string;
@@ -44,46 +45,18 @@ export const TextArea = forwardRef<HTMLTextAreaElement, Props>(
         const [textAreaFocused, setTextAreaFocused] = useState(false);
         const internalRef = useRef<HTMLTextAreaElement>(null);
         const textAreaRef = (ref as RefObject<HTMLTextAreaElement>) || internalRef;
-        const textAreaContentRows = (restProps.value || "").split("\n").length;
 
-        let textAreaRows: number = textAreaContentRows;
+        const { textAreaRows } = useExpandingTextArea(textAreaRef, rows, restProps.value, autoExpand, textAreaFocused);
 
-        if (textAreaFocused && rows > textAreaContentRows) {
-            textAreaRows = rows;
+        function onFocus(e: FocusEvent<HTMLTextAreaElement>) {
+            setTextAreaFocused(true);
+            restProps.onFocus && restProps.onFocus(e);
         }
 
-        useEffect(() => {
-            if (textAreaRef.current && !autoExpand) {
-                // reset style attribute if no longer autoExpanding.
-                textAreaRef.current.setAttribute("style", "");
-            }
-        }, [autoExpand]);
-
-        useEffect(() => {
-            const textAreaElement = textAreaRef.current;
-            if (autoExpand && textAreaElement && !textAreaElement.style.height) {
-                textAreaElement.style.overflowY = "hidden";
-                calculateAndSetElementHeight(textAreaRows, textAreaElement);
-            }
-        }, []);
-
-        useEffect(() => {
-            const textAreaElement = textAreaRef.current;
-            let rowToCalculateHeightFrom = textAreaRows;
-
-            if (textAreaFocused) {
-                // While the text-area has focus we keep the height 1 row higher than the actual row count.
-                // If we don't do this it creates a weird scrolling effect that while we are transitioning from
-                // one height to the next height.
-                rowToCalculateHeightFrom += 1;
-            } else if (restProps.value && rowToCalculateHeightFrom < rows) {
-                rowToCalculateHeightFrom = rows;
-            }
-
-            if (autoExpand && textAreaElement) {
-                calculateAndSetElementHeight(rowToCalculateHeightFrom, textAreaElement);
-            }
-        }, [textAreaRows, textAreaFocused]);
+        function onBlur(e: FocusEvent<HTMLTextAreaElement>) {
+            setTextAreaFocused(false);
+            restProps.onBlur && restProps.onBlur(e);
+        }
 
         return (
             <div data-testid="jkl-text-area" className={componentClassName}>
@@ -110,23 +83,6 @@ export const TextArea = forwardRef<HTMLTextAreaElement, Props>(
                 />
             </div>
         );
-
-        function onFocus(e: FocusEvent<HTMLTextAreaElement>) {
-            setTextAreaFocused(true);
-            restProps.onFocus && restProps.onFocus(e);
-        }
-
-        function onBlur(e: FocusEvent<HTMLTextAreaElement>) {
-            setTextAreaFocused(false);
-            restProps.onBlur && restProps.onBlur(e);
-        }
     },
 );
 TextArea.displayName = "TextArea";
-
-function calculateAndSetElementHeight(rows: number, textAreaElement: HTMLTextAreaElement) {
-    const lineHeightWithPx = window.getComputedStyle(textAreaElement).lineHeight;
-    const lineHeight = parseInt(lineHeightWithPx.replace("px", ""));
-    // 16px is sum of top and bottom padding
-    textAreaElement.style.height = `${rows * lineHeight + 16}px`;
-}
